requirements.txt
python-telegram-bot==22.5
gspread==6.2.1
python-dotenv==1.1.1
pandas==2.3.2